switched from mono to Coral
ScriptEngine:
 - switched for mono to Coral

- const correctness for Buffer
- Added StackBuffer a simple Buffer like interface for fixed size stack buffers
- Hash.GenerateFNV now accepts std.string_view
- restrucured project interface
- The project nolonger storges absolute paths

+++ Shark-Editor/src/Panels/ContentBrowser/ThumbnailGenerator.cpp
@@ -279,7 +279,7 @@ namespace Shark {
 		auto [radianceMap, irradianceMap] = Renderer::CreateEnvironmentMap("Resources/Environment/green_point_park_4k.hdr");
 		skyLight.SceneEnvironment = AssetManager::CreateMemoryOnlyAsset<Environment>(radianceMap, irradianceMap);
 		skyLight.Intensity = 0.8f;
-		skyLight.Lod = AssetManager::GetAsset<Environment>(skyLight.SceneEnvironment)->GetRadianceMap()->GetMipLevelCount() - 1;
+		skyLight.Lod = (float)(AssetManager::GetAsset<Environment>(skyLight.SceneEnvironment)->GetRadianceMap()->GetMipLevelCount() - 1);
 
 		m_CommandBuffer = RenderCommandBuffer::Create();
 	}
